+++ chat-widget/src/components/livechatwidget/common/authHelper.ts
@@ -33,7 +33,15 @@ const handleAuthentication = async (chatSDK: OmnichannelChatSDK, chatConfig: Cha
             TelemetryHelper.logActionEvent(LogLevel.ERROR, { Event: TelemetryEvent.ReceivedNullOrEmptyToken });
             throw new Error(WidgetLoadCustomErrorString.AuthenticationFailedErrorString);
         }
+    } else if (chatSDK?.chatSDKConfig?.getAuthToken) {
+        const token = await chatSDK.chatSDKConfig?.getAuthToken();
+        if (isNullOrEmptyString(token)) {
+            TelemetryHelper.logActionEvent(LogLevel.ERROR, { Event: TelemetryEvent.ReceivedNullOrEmptyToken, Description: "getAuthToken in chat SDK returns empty string" });
+            throw new Error(WidgetLoadCustomErrorString.AuthenticationFailedErrorString);
+        }
+        return { "result": true, token };
     }
+
     return {
         "result": false, 
         "token": null,