+++ chat-widget/src/plugins/newMessageEventHandler.ts
@@ -92,7 +92,7 @@ export const createOnNewAdapterActivityHandler = (chatId: string, userId: string
                     TelemetryHelper.logActionEvent(LogLevel.INFO, {
                         Event: TelemetryEvent.MessageReceived,
                         Description: "New message received",
-                        Data: payload
+                        CustomProperties: payload
                     });
                 } else {
                     if (!isHistoryMessageReceivedEventRasied) {
@@ -100,7 +100,7 @@ export const createOnNewAdapterActivityHandler = (chatId: string, userId: string
                         TelemetryHelper.logActionEvent(LogLevel.INFO, {
                             Event: TelemetryEvent.RehydrateMessageReceived,
                             Description: "History message received",
-                            Data: payload
+                            CustomProperties: payload
                         });
                     }
                 }